seperate to_dot into different uml types
